requirements.txt
blinker==1.9.0
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.59.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
scipy==1.16.1
six==1.17.0
Werkzeug==3.1.3
opencv-python==4.10.0.84
Pillow==11.3.0
PyWavelets==1.4.1
wfdb==4.1.2
